model: move User and Group definition to backend
